// types.ts
export type MapNodeStatus = 'LOCKED' | 'ACTIVE' | 'COMPLETED';

export interface MapLevel {
  id: string;
  title: string;
  description: string;
  xpReward: number;
  order: number; // 1 to 12
  iconName: string; // String reference to Lucide icon
  criteria: string[]; // List of requirements to unlock
  status: MapNodeStatus;
  isTreasure?: boolean;
}
// ...
export type UserRole = 'ADMIN' | 'USER';
export type UserStatus = 'PENDING' | 'ACTIVE' | 'BLOCKED';
export type UserPlan = 'BASIC' | 'PREMIUM';

export interface UserAccount {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  plan: UserPlan; // Added plan field
  joinDate: string;
  
  // Progress Data Snapshot
  currentConsistencyLevel: string; // e.g. "FOCO", "ELITE"
  streak: number;
  avgDiscipline: number;
  lastLogin: string;
}